working cross-compile for macOS x86-64
It works but does not play well with the original cross-compilation
option of the script.

We follow a different path and we do not use the meson cross compilation
because we need to compile also the dependency libraries so we set instead
the variables CC, CXX, CFLAGS etc to the appropriate values and we use
an ordinary meson build.

The build.sh and package.sh should be therefore simplified in such a way
they just get an --arch option used just to name the build directory and
the final package with the architecture name.