fix: problem with reloading doc too quickly
This problem fix a problem occurring sometimes when a file change of
content.

What happens is that when the file changes the OS send a signal to the
directory watch to say the file changed. In turns the application
immediately update the document's content by reading the file. The
problem is sometimes the file is initially truncated and the content is
added afterward. Currently the application can reload the file content
too quickly and get an empty file because the content was added only
later.

With this fix we install a watcher thread for the documents that checks
periodically for changes in the file content for a period of
approximatively two seconds. In the normal case it will load the
document's content only once but if required it will update it multiple
times as soon as changes are observed.