More deterministic file's change watching timing
We compute wake-up time for checking changes in the file's content
changes by using the actual time.

This will be more accurate in case like reading takes a lot of time
like for files in a network filesystem of very large files.
In such case this change reduces also the number of file's lectures.