feat(wip): implement multi-surfaces drawing system
Current status:
Basically work but the mechanism to keep trace of surfaces is flawed.
Some drawing operations are off by one pixel due to rounding. The
DocView tiling drawing is also probably flawed.

New system:

Now every view can create one or more surfaces to draw its content.
Every surface include a RenCache, an SDL_Surface and an SDL_Texture
and is self-contained. The view create surfaces to be drawn at a
given coordinate and later request to the renderer to "present" the
surfaces.

For each surface to "present" means:

- ending the RenCache frame
- Rencache update the rects into the surface pixels
- the whole texture is updated with the surface content
- the texture is drawn in the window using SDL_RenderPresent

Beside this mechanism we now use SDL_RenderSetClipRect from
core.push/pop_clip_rect to limit the drawing zone of each node or view.
The command to set a clip region in the Rencache remains but it is less
used, only to delimit the drawing region within a surface.