feat: we associate an (x, y) origin to each surface
It is done to ease the tiling. The surfaces associated with tiles gets the
(x, y) coordinate of the origin of the tile. In turns, every draw operation
is calculated relatively to this origin. In this way the system can draw
on the surfaces of a tiling system like if they where in the same
coordinate system.