wip: simplify the graphics by embedding destination x, y in surfaces
I make a step-back to take another approach: now we store the x and y
screen coordinates where the surface should be drawn in the RenSurface
itself. In this way we can issue draw operations in the screen coordinates
and the RenSurface will automatically adjust the coordinates to make them
relative to the surface.

In this way we can keep all the old, proven, code performing drawing
operations in screen coordinates, transparently and this greatly simplifies
the porting of the code and reduce the chance of introducing errors.