wip: redraw docview tiles only when needed
Keep trace of the docview's document state to detect changes. If there are
no changes we reuse the previously rendered tiles without doing any drawing
operation on it. We just use the rendered surface as it was previously
drawn.

For the moment does not play well with gutter line's highlight, text
selection and bracketmatch plugin.

There is also a bug: on the first document change, the tile affected is
completely updated in the whole area instead of having the changes
localized in a subset of its cells.