wip: fix selection drawing in docview
We no longer use defer_draw but instead we proceed to draw, inside the
draw() function in multiple-passes. We do in the order:

- draw the tiles for the gutter and the document's content
- present these surfaces
- draw the overlay elements like caret, line highlight, selection
- draw the scrollabar
- present the additional surfaces
- clear unused tiles

The selected text is drawn in the overlay pass drawing a surface for
each line. That works and is quite fine but using a full RenCache for a
small surface that never gets incremental changes is sort of overkill
and unoptimal.