wip: clear rencache's command list on begin_frame
The reason of this change is that we the new drawing architecture we may
give drawing commands to surfaces where we did not call previously
begin_frame. These commands are meant to be discarded as end_frame will
not be called. By resetting command list on begin_frame we ensure these
previous spurious commands are actually discarded.